App-19-non-intrusive-scanning (#57)
