+++ packages/smooth_app/.gitignore
@@ -50,3 +50,4 @@ app.*.map.json
 google-services.json
 GoogleService-Info.plist
 firebase_options.dart
+macos/firebase_app_id_file.json
