+++ packages/smooth_app/lib/pages/scan/scan_page.dart
@@ -1,7 +1,7 @@
 import 'dart:io';
-import 'dart:math' as math;
 
 import 'package:audioplayers/audioplayers.dart';
+import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/rendering.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
@@ -9,12 +9,7 @@ import 'package:provider/provider.dart';
 import 'package:smooth_app/data_models/continuous_scan_model.dart';
 import 'package:smooth_app/data_models/preferences/user_preferences.dart';
 import 'package:smooth_app/data_models/user_management_provider.dart';
-import 'package:smooth_app/generic_lib/design_constants.dart';
-import 'package:smooth_app/generic_lib/dialogs/smooth_alert_dialog.dart';
-import 'package:smooth_app/generic_lib/widgets/smooth_card.dart';
-import 'package:smooth_app/helpers/camera_helper.dart';
 import 'package:smooth_app/helpers/haptic_feedback_helper.dart';
-import 'package:smooth_app/helpers/permission_helper.dart';
 import 'package:smooth_app/pages/scan/camera_scan_page.dart';
 import 'package:smooth_app/widgets/smooth_product_carousel.dart';
 import 'package:smooth_app/widgets/smooth_scaffold.dart';
@@ -33,9 +28,6 @@ class _ScanPageState extends State<ScanPage> {
   AudioPlayer? _musicPlayer;
 
   late UserPreferences _userPreferences;
-
-  /// Percentage of the bottom part of the screen that hosts the carousel.
-  static const int _carouselHeightPct = 55;
 
   @override
   void didChangeDependencies() {
@@ -63,7 +55,6 @@ class _ScanPageState extends State<ScanPage> {
 
     final AppLocalizations appLocalizations = AppLocalizations.of(context);
     final TextDirection direction = Directionality.of(context);
-    final bool hasACamera = CameraHelper.hasACamera;
 
     return SmoothScaffold(
       brightness:
@@ -75,82 +66,70 @@ class _ScanPageState extends State<ScanPage> {
         child: SafeArea(
           child: Container(
             color: Theme.of(context).colorScheme.background,
-            child: Column(
-              children: <Widget>[
-                if (hasACamera)
-                  Expanded(
-                    flex: 100 - _carouselHeightPct,
-                    child: Consumer<PermissionListener>(
-                      builder: (
-                        BuildContext context,
-                        PermissionListener listener,
-                        _,
-                      ) {
-                        switch (listener.value.status) {
-                          case DevicePermissionStatus.checking:
-                            return EMPTY_WIDGET;
-                          case DevicePermissionStatus.granted:
-                            // TODO(m123): change
-                            return const CameraScannerPage();
-                          default:
-                            return const _PermissionDeniedCard();
-                        }
-                      },
-                    ),
-                  ),
-                Expanded(
-                  flex: _carouselHeightPct,
-                  child: Padding(
-                    padding: const EdgeInsetsDirectional.only(bottom: 10.0),
-                    child: SmoothProductCarousel(
-                      containSearchCard: true,
-                      onPageChangedTo: (int page, String? barcode) async {
-                        if (barcode == null) {
-                          // We only notify for new products
-                          return;
-                        }
+            child: Center(
+              child: SizedBox(
+                width: double.infinity,
+                height: MediaQuery.of(context).size.height,
+                child: Padding(
+                  padding: const EdgeInsetsDirectional.only(bottom: 300),
+                  child: SmoothProductCarousel(
+                    containSearchCard: true,
+                    onPageChangedTo: (int page, String? barcode) async {
+                      if (barcode == null) {
+                        // We only notify for new products
+                        return;
+                      }
 
-                        // Both are Future methods, but it doesn't matter to wait here
-                        SmoothHapticFeedback.lightNotification();
+                      // Both are Future methods, but it doesn't matter to wait here
+                      SmoothHapticFeedback.lightNotification();
 
-                        if (_userPreferences.playCameraSound) {
-                          await _initSoundManagerIfNecessary();
-                          await _musicPlayer!.stop();
-                          await _musicPlayer!.play(
-                            AssetSource('audio/beep.wav'),
-                            volume: 0.5,
-                            ctx: const AudioContext(
-                              android: AudioContextAndroid(
-                                isSpeakerphoneOn: false,
-                                stayAwake: false,
-                                contentType: AndroidContentType.sonification,
-                                usageType: AndroidUsageType.notification,
-                                audioFocus:
-                                    AndroidAudioFocus.gainTransientMayDuck,
-                              ),
-                              iOS: AudioContextIOS(
-                                category: AVAudioSessionCategory.soloAmbient,
-                                options: <AVAudioSessionOptions>[
-                                  AVAudioSessionOptions.mixWithOthers,
-                                ],
-                              ),
+                      if (_userPreferences.playCameraSound) {
+                        await _initSoundManagerIfNecessary();
+                        await _musicPlayer!.stop();
+                        await _musicPlayer!.play(
+                          AssetSource('audio/beep.wav'),
+                          volume: 0.5,
+                          ctx: const AudioContext(
+                            android: AudioContextAndroid(
+                              isSpeakerphoneOn: false,
+                              stayAwake: false,
+                              contentType: AndroidContentType.sonification,
+                              usageType: AndroidUsageType.notification,
+                              audioFocus:
+                                  AndroidAudioFocus.gainTransientMayDuck,
                             ),
-                          );
-                        }
-
-                        SemanticsService.announce(
-                          appLocalizations.scan_announce_new_barcode(barcode),
-                          direction,
-                          assertiveness: Assertiveness.assertive,
+                            iOS: AudioContextIOS(
+                              category: AVAudioSessionCategory.soloAmbient,
+                              options: <AVAudioSessionOptions>[
+                                AVAudioSessionOptions.mixWithOthers,
+                              ],
+                            ),
+                          ),
                         );
-                      },
-                    ),
+                      }
+
+                      SemanticsService.announce(
+                        appLocalizations.scan_announce_new_barcode(barcode),
+                        direction,
+                        assertiveness: Assertiveness.assertive,
+                      );
+                    },
                   ),
                 ),
-              ],
+              ),
             ),
           ),
         ),
+      ),
+      floatingActionButton: FloatingActionButton.extended(
+        icon: const Icon(CupertinoIcons.barcode),
+        label: Text(appLocalizations.scan_navbar_label),
+        onPressed: () {
+          Navigator.push(
+              context,
+              MaterialPageRoute<Widget>(
+                  builder: (BuildContext context) => ScannerPage()));
+        },
       ),
     );
   }
@@ -200,114 +179,5 @@ class _ScanPageState extends State<ScanPage> {
   void dispose() {
     _disposeSoundManager();
     super.dispose();
-  }
-}
-
-class _PermissionDeniedCard extends StatelessWidget {
-  const _PermissionDeniedCard({Key? key}) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    final AppLocalizations localizations = AppLocalizations.of(context);
-
-    return SafeArea(
-      child: LayoutBuilder(
-        builder: (BuildContext context, BoxConstraints constraints) {
-          return Container(
-            alignment: Alignment.topCenter,
-            constraints: BoxConstraints.tightForFinite(
-              width: constraints.maxWidth,
-              height: math.min(constraints.maxHeight * 0.9, 200),
-            ),
-            child: SmoothCard(
-              padding: const EdgeInsetsDirectional.only(
-                top: 10.0,
-                start: SMALL_SPACE,
-                end: SMALL_SPACE,
-                bottom: 5.0,
-              ),
-              child: Align(
-                alignment: Alignment.topCenter,
-                child: Column(
-                  children: <Widget>[
-                    Text(
-                      localizations.permission_photo_denied_title,
-                      style: const TextStyle(
-                        fontSize: 18.0,
-                        fontWeight: FontWeight.bold,
-                      ),
-                      textAlign: TextAlign.center,
-                    ),
-                    Expanded(
-                      child: SingleChildScrollView(
-                        child: Padding(
-                          padding: const EdgeInsets.symmetric(
-                            horizontal: 10.0,
-                            vertical: 10.0,
-                          ),
-                          child: Text(
-                            localizations.permission_photo_denied_message(
-                              APP_NAME,
-                            ),
-                            textAlign: TextAlign.center,
-                            style: const TextStyle(
-                              height: 1.4,
-                              fontSize: 15.5,
-                            ),
-                          ),
-                        ),
-                      ),
-                    ),
-                    SmoothActionButtonsBar.single(
-                      action: SmoothActionButton(
-                        text: localizations.permission_photo_denied_button,
-                        onPressed: () => _askPermission(context),
-                      ),
-                    ),
-                  ],
-                ),
-              ),
-            ),
-          );
-        },
-      ),
-    );
-  }
-
-  Future<void> _askPermission(BuildContext context) {
-    return Provider.of<PermissionListener>(
-      context,
-      listen: false,
-    ).askPermission(onRationaleNotAvailable: () async {
-      return showDialog(
-          context: context,
-          builder: (BuildContext context) {
-            final AppLocalizations localizations = AppLocalizations.of(context);
-
-            return SmoothAlertDialog(
-              title:
-                  localizations.permission_photo_denied_dialog_settings_title,
-              body: Text(
-                localizations.permission_photo_denied_dialog_settings_message,
-                style: const TextStyle(
-                  height: 1.6,
-                ),
-              ),
-              negativeAction: SmoothActionButton(
-                text: localizations
-                    .permission_photo_denied_dialog_settings_button_cancel,
-                onPressed: () => Navigator.of(context).pop(false),
-                lines: 2,
-              ),
-              positiveAction: SmoothActionButton(
-                text: localizations
-                    .permission_photo_denied_dialog_settings_button_open,
-                onPressed: () => Navigator.of(context).pop(true),
-                lines: 2,
-              ),
-              actionsAxis: Axis.vertical,
-            );
-          });
-    });
   }
 }
